use purser-metamask as identity provider
